+++ src/Problems/Rocket/ProblemRocket_Terrain.h
@@ -22,30 +22,30 @@ protected:
 	// World specs
 	int terrainMaxHeight;
 	int terrainMinHeight;
-	double windHorizontal;
-	double windVertical;
-	double gravity;
+	float windHorizontal;
+	float windVertical;
+	float gravity;
 
 	// LandingZone
 	int terrainFlatZone1;
 	int terrainFlatZone2;
 
 	// Rocket start
-	double rocketStartX;
-	double rocketStartY;
-	double rocketStartHSpeed;
-	double rocketStartVSpeed;
+	float rocketStartX;
+	float rocketStartY;
+	float rocketStartHSpeed;
+	float rocketStartVSpeed;
 
 	// Terrain points
 	vector<pair<int, int>> terrain;	// The terrain. Just specify points, the collision and drawing is automatic
 
 	// Helper functions
-	vector<pair<int, int>>::iterator getPointBefore(double x);	// Returns the point contained in this->terrain with closest smallest x compared to given x parameter
-	double getTerrainPoint(double x, pair<int, int> p1, pair<int, int> p2);	// Gets the Y coordinate for the given X, with p1.x <=x and p2.x >= x
+	vector<pair<int, int>>::iterator getPointBefore(float x);	// Returns the point contained in this->terrain with closest smallest x compared to given x parameter
+	float getTerrainPoint(float x, pair<int, int> p1, pair<int, int> p2);	// Gets the Y coordinate for the given X, with p1.x <=x and p2.x >= x
 
 public:
 	// Constructors
-	ProblemRocket_Terrain(int width, int height, int maxTerrainHeight, int minTerrainHeight, double rocketStartX, double rocketStartY, double rocketStartHSPeed, double rocketStartVSpeed, double gravity = 4);
+	ProblemRocket_Terrain(int width, int height, int maxTerrainHeight, int minTerrainHeight, float rocketStartX, float rocketStartY, float rocketStartHSPeed, float rocketStartVSpeed, float gravity = 4);
 	ProblemRocket_Terrain(int width, int height);	// easy constructor
 	ProblemRocket_Terrain(string path = "Default");	// Base constructor
 
@@ -58,19 +58,19 @@ public:
 	virtual void generateRandomTerrain(int width = -1, int height = -1);	// Leave -1, -1 to keep the current size
 
 	// Terrain interaction
-	bool isOnMap(double x, double y);
-	bool collides(double x, double y);	// Checks if the given position is colliding with terrain
-	double getTerrainPoint(double x);	// Gets the Y coordinate for the given X, with p1.x <=x and p2.x >= x
+	bool isOnMap(float x, float y);
+	bool collides(float x, float y);	// Checks if the given position is colliding with terrain
+	float getTerrainPoint(float x);	// Gets the Y coordinate for the given X, with p1.x <=x and p2.x >= x
 
 	// Getters
 	vector<pair<int, int>>* getTerrain();
-	void getLandingZone(double& landing1, double& landing2);
-	void getTerrainDimensions(double& terrainWidth, double& terrainHeight);
+	void getLandingZone(float& landing1, float& landing2);
+	void getTerrainDimensions(float& terrainWidth, float& terrainHeight);
 	int getWidth();
 	int getHeight();
-	void getRocketStart(double& rocketX, double& rocketY, double& rocketHorizontalSpeed, double& rocketVerticalSpeed);
-	void getWind(double& windHorizontal, double& windVertical);
-	double getGravity();
+	void getRocketStart(float& rocketX, float& rocketY, float& rocketHorizontalSpeed, float& rocketVerticalSpeed);
+	void getWind(float& windHorizontal, float& windVertical);
+	float getGravity();
 };
 
 #endif
