+++ src/Utilities/MathHelper.h
@@ -2,12 +2,13 @@
 #define MATHHELPER_H_
 
 #include <math.h>
+#include <iostream>
 
 #define PI (3.141592653589793)
 
-double degToRad(double degrees);
-double radToDeg(double radians);
-double computeAngle(double x, double y);	// Computes angle (in degrees) from given
+float degToRad(float degrees);
+float radToDeg(float radians);
+float computeAngle(float x, float y);	// Computes angle (in degrees) from given
 											// Helper functions
 float convertToRange(float value, float minValue, float maxValue, float targetMinValue, float targetMaxValue);	// Converts given value from its range to the given range
